api/pyproject.toml
[tool.poetry]
name = "octoml-sample-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.78.0"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
onnxruntime = "^1.11.1"
requests = "^2.28.0"
python-multipart = "^0.0.5"
tritonclient = {extras = ["http", "grpc"], version = "^2.22.3"}
Pillow = "^9.1.1"
attrdict = "^2.0.1"
numpy = "^1.22.4"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pytest = "^7.1.2"
flake8 = "^4.0.1"
pylint = "^2.14.1"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
